Split up css, helper and View into more files
