+++ frontend/src/assets/stylesheets/cms-view.css
@@ -1,0 +1,117 @@
+/*
+ * CMS view-level styles
+ * Owns: modal, side panel and create-form presentation.
+ * Avoid adding AG Grid internals here; use cms-ag-grid.css for that.
+ */
+
+.cms-add-button {
+  @apply inline-flex w-fit items-center rounded-full border border-transparent bg-surface-inv px-5 py-2.5 text-sm font-semibold text-ink-on-inv transition hover:bg-surface-inv-raised;
+}
+
+.cms-side-panel {
+  @apply fixed right-0 top-0 z-50 flex h-screen w-full max-w-xl flex-col border-l border-surface-3 bg-surface-0 p-5 shadow-2xl;
+}
+
+.cms-side-header {
+  @apply mb-4 flex items-center justify-between;
+}
+
+.cms-side-close {
+  @apply rounded-md border border-surface-3 px-3 py-1.5 text-sm text-ink-secondary transition hover:bg-surface-1;
+}
+
+.cms-side-body {
+  @apply flex flex-1 flex-col gap-4 overflow-y-auto;
+}
+
+.cms-side-field {
+  @apply flex flex-col gap-2;
+}
+
+.cms-side-textarea {
+  @apply min-h-28 rounded-md border border-surface-3 bg-surface-1 p-3 text-sm text-ink-primary;
+}
+
+.cms-side-footer {
+  @apply mt-4 flex justify-end;
+}
+
+.cms-side-save {
+  @apply rounded-md bg-surface-inv px-4 py-2 text-sm font-semibold text-ink-on-inv transition hover:bg-surface-inv-raised disabled:cursor-not-allowed disabled:opacity-60;
+}
+
+.cms-modal-overlay {
+  @apply fixed inset-0 z-50 flex items-center justify-center bg-black/40 p-4;
+}
+
+.cms-modal {
+  @apply flex h-[85vh] w-full max-w-5xl flex-col overflow-hidden rounded-xl border border-surface-3 bg-surface-0;
+}
+
+.cms-modal-header {
+  @apply flex items-center justify-between border-b border-surface-3 px-5 py-4;
+}
+
+.cms-modal-body {
+  @apply flex-1 space-y-5 overflow-y-auto px-5 py-4;
+}
+
+.cms-modal-footer {
+  @apply flex justify-end gap-2 border-t border-surface-3 px-5 py-4;
+}
+
+.cms-form-block {
+  @apply rounded-lg border border-surface-3 bg-surface-1 p-4;
+}
+
+.cms-form-legend {
+  @apply px-1 text-sm font-semibold text-ink-primary;
+}
+
+.cms-required {
+  @apply ml-1 text-red-600;
+}
+
+.cms-lang-grid {
+  @apply mt-3 grid grid-cols-1 gap-3 md:grid-cols-3;
+}
+
+.cms-lang-grid-single {
+  @apply grid-cols-1 md:grid-cols-1;
+}
+
+.cms-lang-grid-double {
+  @apply grid-cols-1 md:grid-cols-2;
+}
+
+.cms-form-lang-field {
+  @apply flex flex-col gap-2;
+}
+
+.cms-language-toggle-row {
+  @apply flex flex-wrap items-center justify-between gap-3 rounded-md border border-surface-3 bg-surface-0 px-3 py-2;
+}
+
+.cms-language-pill {
+  @apply rounded-full border border-surface-3 px-3 py-1 text-xs font-semibold text-ink-secondary transition hover:bg-surface-1;
+}
+
+.cms-language-pill.active {
+  @apply border-transparent bg-surface-inv text-ink-on-inv;
+}
+
+.cms-upload-controls {
+  @apply grid grid-cols-1 gap-3 md:grid-cols-3;
+}
+
+.cms-lang-label {
+  @apply text-xs font-semibold uppercase tracking-wide text-ink-secondary;
+}
+
+.cms-text-input {
+  @apply rounded-md border border-surface-3 bg-surface-0 px-3 py-2 text-sm text-ink-primary;
+}
+
+.cms-toggle-row {
+  @apply flex items-center gap-2 text-sm text-ink-primary;
+}
