+++ frontend/src/i18n/locales/en.ts
@@ -67,6 +67,13 @@ export default {
       searchPlaceholder: "Search across all columns...",
       selectedCount: "Selected: {count}",
       loadedCount: "Rows loaded: {count}",
+      fitColumns: "Fit columns",
+      autoSize: "Auto-size",
+      resetFilters: "Reset filters",
+      exportCsv: "Export CSV",
+      resetState: "Reset state",
+      columns: "Columns",
+      hideColumns: "Hide columns",
       noRows: "No productions found for the current filters.",
       selectAll: "Select all",
       clearSelection: "Clear selection",
