+++ frontend/src/i18n/locales/fr.ts
@@ -67,6 +67,13 @@ export default {
       searchPlaceholder: "Rechercher dans toutes les colonnes...",
       selectedCount: "Selectionne: {count}",
       loadedCount: "Lignes chargees: {count}",
+      fitColumns: "Ajuster les colonnes",
+      autoSize: "Taille automatique",
+      resetFilters: "Reinitialiser les filtres",
+      exportCsv: "Exporter CSV",
+      resetState: "Reinitialiser l'etat",
+      columns: "Colonnes",
+      hideColumns: "Masquer les colonnes",
       noRows: "Aucune production trouvee pour les filtres actuels.",
       selectAll: "Tout selectionner",
       clearSelection: "Effacer la selection",
