+++ frontend/src/i18n/locales/nl.ts
@@ -67,6 +67,13 @@ export default {
       searchPlaceholder: "Zoek in alle kolommen...",
       selectedCount: "Geselecteerd: {count}",
       loadedCount: "Rijen geladen: {count}",
+      fitColumns: "Kolommen passend maken",
+      autoSize: "Automatisch schalen",
+      resetFilters: "Filters resetten",
+      exportCsv: "CSV exporteren",
+      resetState: "Status resetten",
+      columns: "Kolommen",
+      hideColumns: "Kolommen verbergen",
       noRows: "Geen producties gevonden voor de huidige filters.",
       selectAll: "Selecteer alles",
       clearSelection: "Wis selectie",
