+++ frontend/src/services/cms/date.ts
@@ -1,0 +1,13 @@
+export function toLocalDateTimeInput(value: Date | string): string {
+  const date = new Date(value);
+  if (Number.isNaN(date.getTime())) {
+    return "";
+  }
+
+  const local = new Date(date.getTime() - date.getTimezoneOffset() * 60_000);
+  return local.toISOString().slice(0, 16);
+}
+
+export function toIsoStringFromLocalInput(value: string): string {
+  return new Date(value).toISOString();
+}
